requirements.txt
Django==1.9
djangorestframework==3.3.0
django-rest-swagger==0.3.10
